+++ src/router/routes.js
@@ -1,5 +1,3 @@
-import { createRouter, createWebHistory } from 'vue-router'
-
 export const routes = [
 
 	{
@@ -77,31 +75,4 @@ export const routes = [
 ];
 
 
-
-export const router = createRouter({
-	history: createWebHistory(),
-	routes
-  });
   
-  router.beforeEach((to, from, next) => {
-	const token = localStorage.buscarItemLocalStorage('token');
-  
-	if (to.name === 'Login') {
-	  if (token) {
-		next({ name: 'DashboardView' });
-	  } else {
-		next();
-	  }
-	} else if (to.matched.some(route => route.meta.requiresAuth)) {
-	  if (!token) {
-		next({
-		  path: '/login',
-		  params: { nextUrl: to.fullPath },
-		});
-	  } else {
-		next();
-	  }
-	} else {
-	  next();
-	}
-  });
